a little Wait before initialize
For some reason if the initialization is executed instantly after loading, it doesn't load, I don't know why but at least it happens to me, it's weird but it doesn't receive the source correctly, with a simple wait the problem was solved and the interface loads perfectly. For those who have that problem, it should work for them.